Add exception handling to service classes

+++ src/main/java/com/alatoo/socialEventManagement/services/comment/CommentServiceJPA.java
@@ -1,5 +1,6 @@
 package com.alatoo.socialEventManagement.services.comment;
 
+import com.alatoo.socialEventManagement.controllers.exceptions.NotFoundException;
 import com.alatoo.socialEventManagement.dto.CommentDTO;
 import com.alatoo.socialEventManagement.entities.Comment;
 import com.alatoo.socialEventManagement.mappers.CommentMapper;
@@ -11,7 +12,7 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
-public class CommentServiceJPA implements CommentService{
+public class CommentServiceJPA implements CommentService {
     private final CommentRepository commentRepository;
     private final CommentMapper commentMapper;
 
@@ -22,8 +23,8 @@ public class CommentServiceJPA implements CommentService{
 
     @Override
     public List<CommentDTO> findAllComments() {
-        List<Comment> events = (List<Comment>) commentRepository.findAll();
-        return events.stream()
+        List<Comment> comments = (List<Comment>) commentRepository.findAll();
+        return comments.stream()
                 .map(commentMapper::commentToCommentDto)
                 .collect(Collectors.toList());
     }
@@ -31,9 +32,8 @@ public class CommentServiceJPA implements CommentService{
     @Override
     public Optional<CommentDTO> findCommentById(Long id) {
         Optional<Comment> optionalComment = commentRepository.findById(id);
-        return Optional.ofNullable(
-                commentMapper.commentToCommentDto(optionalComment.orElse(null))
-        );
+        Comment comment = optionalComment.orElseThrow(() -> new NotFoundException("Comment not found with id: " + id));
+        return Optional.of(commentMapper.commentToCommentDto(comment));
     }
 
     @Override
@@ -44,6 +44,9 @@ public class CommentServiceJPA implements CommentService{
 
     @Override
     public void deleteComment(Long id) {
+        if (!commentRepository.existsById(id)) {
+            throw new NotFoundException("Comment not found with id: " + id);
+        }
         commentRepository.deleteById(id);
     }
 }
